+++ src/main/java/roomescape/controller/api/AuthController.java
@@ -7,11 +7,11 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.request.LoginRequest;
-import roomescape.dto.response.MemberResponse;
+import roomescape.controller.dto.request.LoginRequest;
 import roomescape.security.exception.UnauthorizedException;
 import roomescape.service.AuthService;
 import roomescape.service.MemberService;
+import roomescape.service.dto.response.MemberResponse;
 import roomescape.util.CookieUtil;
 
 @RestController
