+++ src/main/java/roomescape/controller/api/MemberController.java
@@ -9,10 +9,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.controller.dto.request.SignupRequest;
 import roomescape.controller.dto.response.ApiResponses;
-import roomescape.dto.request.SignupRequest;
-import roomescape.dto.response.MemberResponse;
 import roomescape.service.MemberService;
+import roomescape.service.dto.response.MemberResponse;
 
 @RestController
 @RequestMapping("/members")
