+++ src/main/java/roomescape/controller/api/ReservationController.java
@@ -15,13 +15,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.controller.dto.request.ReservationRequest;
 import roomescape.controller.dto.response.ApiResponses;
 import roomescape.controller.support.Auth;
-import roomescape.dto.request.ReservationRequest;
-import roomescape.dto.response.PersonalReservationResponse;
-import roomescape.dto.response.ReservationResponse;
 import roomescape.security.Accessor;
 import roomescape.service.ReservationService;
+import roomescape.service.dto.response.PersonalReservationResponse;
+import roomescape.service.dto.response.ReservationResponse;
 
 @RestController
 @RequestMapping("/reservations")
