+++ src/main/java/roomescape/controller/api/ThemeController.java
@@ -16,10 +16,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.controller.dto.request.ThemeRequest;
 import roomescape.controller.dto.response.ApiResponses;
-import roomescape.dto.request.ThemeRequest;
-import roomescape.dto.response.ThemeResponse;
 import roomescape.service.ThemeService;
+import roomescape.service.dto.response.ThemeResponse;
 
 @RestController
 @RequestMapping("/themes")
