+++ src/main/java/roomescape/service/ReservationService.java
@@ -20,9 +20,9 @@ import roomescape.domain.reservationwaiting.ReservationWaitingRepository;
 import roomescape.domain.reservationwaiting.WaitingWithRank;
 import roomescape.domain.theme.Theme;
 import roomescape.domain.theme.ThemeRepository;
-import roomescape.dto.response.PersonalReservationResponse;
-import roomescape.dto.response.ReservationResponse;
-import roomescape.service.dto.CreateReservationRequest;
+import roomescape.service.dto.request.CreateReservationRequest;
+import roomescape.service.dto.response.PersonalReservationResponse;
+import roomescape.service.dto.response.ReservationResponse;
 
 @Service
 @Transactional(readOnly = true)
