+++ src/main/java/roomescape/service/dto/request/CreateReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.dto.request;
 
 import java.time.LocalDate;
 import roomescape.domain.member.Member;
