+++ src/main/java/roomescape/service/dto/response/MemberResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto.response;
+package roomescape.service.dto.response;
 
 import roomescape.domain.member.Member;
 import roomescape.domain.member.Role;
