+++ src/main/java/roomescape/service/dto/response/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto.response;
+package roomescape.service.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
