+++ src/test/java/roomescape/service/AuthServiceTest.java
@@ -11,11 +11,11 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.SpyBean;
+import roomescape.controller.dto.request.LoginRequest;
+import roomescape.controller.dto.request.SignupRequest;
 import roomescape.domain.member.Role;
-import roomescape.dto.request.LoginRequest;
-import roomescape.dto.request.SignupRequest;
-import roomescape.dto.response.MemberResponse;
 import roomescape.security.TokenProvider;
+import roomescape.service.dto.response.MemberResponse;
 
 class AuthServiceTest extends BaseServiceTest {
