+++ src/test/java/roomescape/service/ReservationWaitingServiceTest.java
@@ -19,8 +19,8 @@ import roomescape.domain.reservationwaiting.ReservationWaiting;
 import roomescape.domain.reservationwaiting.ReservationWaitingRepository;
 import roomescape.domain.theme.Theme;
 import roomescape.domain.theme.ThemeRepository;
-import roomescape.dto.response.ReservationResponse;
-import roomescape.service.dto.CreateReservationRequest;
+import roomescape.service.dto.request.CreateReservationRequest;
+import roomescape.service.dto.response.ReservationResponse;
 import roomescape.support.fixture.MemberFixture;
 import roomescape.support.fixture.ReservationFixture;
 import roomescape.support.fixture.ReservationTimeFixture;
